refactor(st-09072): modularize insert executor
