+++ docs/st09072-relational-insert-executor-modularization.md
@@ -1,0 +1,44 @@
+# ST-09072: Relational Insert Executor Modularization
+
+## Summary
+
+`packages/tools/src/data/relational/tools/relational-insert/executor.ts` was reduced from a `365` line mixed-responsibility runtime to a `91` line public facade. The insert execution flow now delegates to focused internal helpers for result normalization and option resolution, single-query execution, and batch orchestration while preserving the public `executeInsert(...)` surface.
+
+## Production Split
+
+| File | Role | Lines |
+| --- | --- | ---: |
+| `packages/tools/src/data/relational/tools/relational-insert/executor.ts` | Stable public facade, logging, timing, error wrapping | `91` |
+| `packages/tools/src/data/relational/tools/relational-insert/executor-shared.ts` | Result normalization, inserted-id derivation, batch-option defaults, shared logger/context | `133` |
+| `packages/tools/src/data/relational/tools/relational-insert/executor-single.ts` | Single-query and heterogeneous SQLite execution flow | `80` |
+| `packages/tools/src/data/relational/tools/relational-insert/executor-batch.ts` | Batch orchestration, progress logging, synthetic benchmark metadata | `90` |
+
+## Test Split
+
+The old `packages/tools/tests/data/relational/tools/insert-executor.test.ts` monolith was replaced with a public entrypoint plus focused suites:
+
+| File | Role | Lines |
+| --- | --- | ---: |
+| `packages/tools/tests/data/relational/tools/insert-executor.test.ts` | Stable public test entrypoint | `7` |
+| `packages/tools/tests/data/relational/tools/insert-executor/shared.ts` | Shared mock manager and executor import | `11` |
+| `packages/tools/tests/data/relational/tools/insert-executor/result-shaping.suite.ts` | Row-count normalization, returning behavior, inserted-id derivation | `131` |
+| `packages/tools/tests/data/relational/tools/insert-executor/batch-mode.suite.ts` | Batch enablement and default option behavior | `67` |
+| `packages/tools/tests/data/relational/tools/insert-executor/error-handling.suite.ts` | Constraint wrapping, safe validation passthrough, sanitized unknown errors, transaction routing | `72` |
+
+## Test Strategy
+
+This story used a characterization-first approach instead of forcing a red-first structural test. The change is intentionally behavior-preserving and primarily reorganizes responsibilities, so an initial failing test would mostly assert temporary file layout rather than the stable `executeInsert(...)` contract. The practical safety net was to split the executor coverage first, keep a public entrypoint, and run the focused executor and insert-domain surfaces before and after the production refactor.
+
+## Validation
+
+- `pnpm test --run packages/tools/tests/data/relational/tools/insert-executor.test.ts` -> `17` passed
+- `pnpm test --run packages/tools/tests/data/relational/relational-insert/index.test.ts` -> `13` passed, `13` skipped
+- `pnpm test --run packages/tools/tests/data/relational/tools/insert-tool.test.ts` -> `5` passed
+- `pnpm --filter @agentforge/tools typecheck` -> passed
+- `pnpm lint:explicit-any:baseline` -> passed at `workspace 80/289`, `tools 53/67`
+
+## Compatibility Notes
+
+- `executeInsert(...)` remains the stable public entrypoint.
+- Batch semantics, synthetic benchmark metadata, transaction-context routing, constraint wrapping, and inserted-id derivation remain unchanged.
+- No CI or release automation changes were required because the refactor stays behind the existing tool/runtime interfaces and validation commands.
