+++ packages/tools/src/data/relational/tools/relational-insert/executor.ts
@@ -3,296 +3,22 @@
  * @module tools/relational-insert/executor
  */
 
-import { createLogger } from '@agentforge/core';
-import { buildInsertQuery } from '../../query/query-builder.js';
-import {
-  benchmarkBatchExecution,
-  executeBatchedTask,
-} from '../../query/batch-executor.js';
 import type { ConnectionManager } from '../../connection/connection-manager.js';
-import type { TransactionContext } from '../../query/transaction.js';
 import type {
-  InsertRow,
-  InsertBatchMetadata,
-  InsertBatchOptions,
   InsertResult,
+  InsertRow,
   RelationalInsertInput,
 } from './types.js';
 import { getConstraintViolationMessage, isSafeInsertValidationError } from './error-utils.js';
+import { executeInsertInBatchMode } from './executor-batch.js';
+import {
+  insertExecutorLogger,
+  resolveBatchOptions,
+  type InsertExecutionContext,
+} from './executor-shared.js';
+import { executeInsertOnce } from './executor-single.js';
 
-const logger = createLogger('agentforge:tools:data:relational:insert');
-
-interface NormalizedExecutionResult {
-  rows: unknown[];
-  rowCount: number;
-  insertId?: number;
-  lastInsertRowid?: number;
-}
-
-/**
- * Execution context for INSERT operations.
- *
- * @property transaction - Optional active transaction to execute within
- */
-export interface InsertExecutionContext {
-  transaction?: TransactionContext;
-}
-
-function toNumber(value: unknown): number | undefined {
-  return typeof value === 'number' && Number.isFinite(value) ? value : undefined;
-}
-
-function isPlainObject(value: unknown): value is Record<string, unknown> {
-  return !!value && typeof value === 'object' && !Array.isArray(value);
-}
-
-function normalizeExecutionResult(result: unknown): NormalizedExecutionResult {
-  if (Array.isArray(result)) {
-    if (result.length > 0 && isPlainObject(result[0])) {
-      const first = result[0];
-      const affectedRows = toNumber(first.affectedRows) ?? toNumber(first.rowCount) ?? toNumber(first.changes);
-      const insertId = toNumber(first.insertId);
-      const lastInsertRowid = toNumber(first.lastInsertRowid);
-
-      if (affectedRows !== undefined || insertId !== undefined || lastInsertRowid !== undefined) {
-        return {
-          rows: [],
-          rowCount: affectedRows ?? 0,
-          insertId,
-          lastInsertRowid,
-        };
-      }
-    }
-
-    return {
-      rows: result,
-      rowCount: result.length,
-    };
-  }
-
-  if (isPlainObject(result)) {
-    const rows = Array.isArray(result.rows) ? result.rows : [];
-    const rowCount = toNumber(result.rowCount)
-      ?? toNumber(result.affectedRows)
-      ?? toNumber(result.changes)
-      ?? rows.length;
-
-    return {
-      rows,
-      rowCount,
-      insertId: toNumber(result.insertId),
-      lastInsertRowid: toNumber(result.lastInsertRowid),
-    };
-  }
-
-  return {
-    rows: [],
-    rowCount: 0,
-  };
-}
-
-function deriveInsertedIds(options: {
-  idColumn: string;
-  inputRows: Array<Record<string, unknown>>;
-  returnedRows: unknown[];
-  rowCount: number;
-  insertId?: number;
-  lastInsertRowid?: number;
-}): Array<number | string> {
-  const { idColumn, inputRows, returnedRows, rowCount, insertId, lastInsertRowid } = options;
-
-  if (returnedRows.length > 0) {
-    const idsFromReturn = returnedRows
-      .map((row) => (isPlainObject(row) ? row[idColumn] : undefined))
-      .filter((id): id is number | string => typeof id === 'number' || typeof id === 'string');
-
-    if (idsFromReturn.length > 0) {
-      return idsFromReturn;
-    }
-  }
-
-  if (inputRows.every((row) => typeof row[idColumn] === 'number' || typeof row[idColumn] === 'string')) {
-    return inputRows.map((row) => row[idColumn] as number | string);
-  }
-
-  if (insertId !== undefined && rowCount > 0) {
-    return Array.from({ length: rowCount }, (_, index) => insertId + index);
-  }
-
-  if (lastInsertRowid !== undefined && rowCount > 0) {
-    const startId = lastInsertRowid - rowCount + 1;
-    return Array.from({ length: rowCount }, (_, index) => startId + index);
-  }
-
-  return [];
-}
-
-function resolveBatchOptions(batch: InsertBatchOptions | undefined): Required<InsertBatchOptions> | undefined {
-  if (!batch || batch.enabled === false) {
-    return undefined;
-  }
-
-  return {
-    enabled: batch.enabled ?? true,
-    batchSize: batch.batchSize ?? 100,
-    continueOnError: batch.continueOnError ?? true,
-    maxRetries: batch.maxRetries ?? 0,
-    retryDelayMs: batch.retryDelayMs ?? 0,
-    benchmark: batch.benchmark ?? false,
-  };
-}
-
-async function executeInsertOnce(
-  manager: ConnectionManager,
-  input: RelationalInsertInput,
-  context?: InsertExecutionContext
-): Promise<InsertResult> {
-  const built = buildInsertQuery({
-    table: input.table,
-    data: input.data,
-    returning: input.returning,
-    vendor: input.vendor,
-  });
-
-  const executor = context?.transaction ?? manager;
-
-  // When the query builder returns SQL[] (heterogeneous SQLite rows),
-  // execute each statement individually and aggregate results.
-  if (Array.isArray(built.query)) {
-    let totalRowCount = 0;
-    const allRows: unknown[] = [];
-    const allIds: Array<string | number> = [];
-
-    for (const singleQuery of built.query) {
-      const rawResult = await executor.execute(singleQuery);
-      const normalized = normalizeExecutionResult(rawResult);
-      totalRowCount += normalized.rowCount > 0 ? normalized.rowCount : 1;
-      if (built.returningMode === 'row') {
-        allRows.push(...normalized.rows);
-      }
-      if (built.returningMode === 'id') {
-        const ids = deriveInsertedIds({
-          idColumn: built.idColumn,
-          inputRows: [built.rows[allIds.length]],
-          returnedRows: normalized.rows,
-          rowCount: 1,
-          insertId: normalized.insertId,
-          lastInsertRowid: normalized.lastInsertRowid,
-        });
-        allIds.push(...ids);
-      }
-    }
-
-    return {
-      rowCount: totalRowCount,
-      rows: allRows,
-      insertedIds: allIds,
-      executionTime: 0,
-    };
-  }
-
-  const rawResult = await executor.execute(built.query);
-  const normalized = normalizeExecutionResult(rawResult);
-
-  const rowCount = normalized.rowCount > 0 ? normalized.rowCount : built.rows.length;
-  const rows = built.returningMode === 'row' ? normalized.rows : [];
-  const insertedIds = built.returningMode === 'id'
-    ? deriveInsertedIds({
-      idColumn: built.idColumn,
-      inputRows: built.rows,
-      returnedRows: normalized.rows,
-      rowCount,
-      insertId: normalized.insertId,
-      lastInsertRowid: normalized.lastInsertRowid,
-    })
-    : [];
-
-  return {
-    rowCount,
-    rows,
-    insertedIds,
-    executionTime: 0,
-  };
-}
-
-async function executeInsertInBatchMode(
-  manager: ConnectionManager,
-  input: RelationalInsertInput & { data: InsertRow[] },
-  context: InsertExecutionContext | undefined,
-  options: Required<InsertBatchOptions>
-): Promise<InsertResult> {
-  const batchResult = await executeBatchedTask<InsertRow, InsertResult>(
-    {
-      operation: 'insert',
-      items: input.data,
-      executeBatch: async (rows) =>
-        executeInsertOnce(manager, {
-          ...input,
-          data: rows,
-          batch: undefined,
-        }, context),
-      getBatchSuccessCount: (result, rows) => Math.min(result.rowCount, rows.length),
-    },
-    {
-      batchSize: options.batchSize,
-      continueOnError: options.continueOnError,
-      maxRetries: options.maxRetries,
-      retryDelayMs: options.retryDelayMs,
-      onProgress: (progress) => {
-        logger.debug('INSERT batch progress', {
-          vendor: input.vendor,
-          table: input.table,
-          ...progress,
-        });
-      },
-    }
-  );
-
-  const rowCount = batchResult.results.reduce((total, result) => total + result.rowCount, 0);
-  const rows = batchResult.results.flatMap((result) => result.rows);
-  const insertedIds = batchResult.results.flatMap((result) => result.insertedIds);
-
-  let benchmark: InsertBatchMetadata['benchmark'];
-  if (options.benchmark) {
-    logger.warn('INSERT batch benchmark enabled. Synthetic benchmark callbacks are side-effect free and do not execute SQL.', {
-      vendor: input.vendor,
-      table: input.table,
-      totalRows: input.data.length,
-      batchSize: options.batchSize,
-    });
-
-    benchmark = await benchmarkBatchExecution({
-      items: input.data,
-      batchSize: options.batchSize,
-      runIndividual: async () => {
-        // Synthetic benchmark callback for timing-only comparison.
-      },
-      runBatch: async () => {
-        // Synthetic benchmark callback for timing-only comparison.
-      },
-    });
-  }
-
-  return {
-    rowCount,
-    rows,
-    insertedIds,
-    executionTime: batchResult.executionTime,
-    batch: {
-      enabled: true,
-      batchSize: options.batchSize,
-      totalItems: batchResult.totalItems,
-      processedItems: batchResult.processedItems,
-      successfulItems: batchResult.successfulItems,
-      failedItems: batchResult.failedItems,
-      totalBatches: batchResult.totalBatches,
-      retries: batchResult.retries,
-      partialSuccess: batchResult.partialSuccess,
-      failures: batchResult.failures,
-      benchmark,
-    },
-  };
-}
+export type { InsertExecutionContext } from './executor-shared.js';
 
 /**
  * Execute an INSERT query using the shared query builder.
@@ -304,7 +30,7 @@ export async function executeInsert(
 ): Promise<InsertResult> {
   const startTime = Date.now();
 
-  logger.debug('Building INSERT query', {
+  insertExecutorLogger.debug('Building INSERT query', {
     vendor: input.vendor,
     table: input.table,
     isBatch: Array.isArray(input.data),
@@ -326,7 +52,7 @@ export async function executeInsert(
 
     const executionTime = Date.now() - startTime;
 
-    logger.debug('INSERT query executed successfully', {
+    insertExecutorLogger.debug('INSERT query executed successfully', {
       vendor: input.vendor,
       table: input.table,
       rowCount: result.rowCount,
@@ -344,7 +70,7 @@ export async function executeInsert(
   } catch (error) {
     const executionTime = Date.now() - startTime;
 
-    logger.error('INSERT query execution failed', {
+    insertExecutorLogger.error('INSERT query execution failed', {
       vendor: input.vendor,
       table: input.table,
       error: error instanceof Error ? error.message : String(error),
