+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-06-24
-**Active Story:** ST-09072 - Modularize Relational Insert Executor and Tests (Ready)
+**Last Updated:** 2026-06-25
+**Active Story:** ST-09072 - Modularize Relational Insert Executor and Tests (In Progress)
 
 ---
